refactor : n+1 문제를 로깅하는 로직을 별도의 클래스로 분리
